chore: move all plugin code in src directories
