+++ plugins/conftest-kubernetes/package.json
@@ -2,7 +2,7 @@
   "name": "@garden-io/garden-conftest-kubernetes",
   "version": "0.13.14",
   "description": "Auto-generator for the conftest plugin and Garden kubernetes/helm modules",
-  "main": "build/index.js",
+  "main": "build/src/index.js",
   "private": true,
   "dependencies": {
     "@garden-io/core": "*",
