+++ plugins/conftest/package.json
@@ -2,7 +2,7 @@
   "name": "@garden-io/garden-conftest",
   "version": "0.13.14",
   "description": "conftest plugin for Garden",
-  "main": "build/index.js",
+  "main": "build/src/index.js",
   "private": true,
   "dependencies": {
     "@garden-io/core": "*",
