+++ plugins/conftest/test/conftest.ts
@@ -12,7 +12,7 @@ import { join } from "path"
 
 import { dedent } from "@garden-io/sdk/build/src/util/string"
 import { defaultNamespace } from "@garden-io/sdk/build/src/constants"
-import { gardenPlugin } from ".."
+import { gardenPlugin } from "../src/index"
 import { ProjectConfig } from "@garden-io/sdk/build/src/types"
 import { makeTestGarden } from "@garden-io/sdk/build/src/testing"
