+++ plugins/jib/src/gradle.ts
