+++ plugins/jib/test/index.ts
@@ -11,8 +11,8 @@ import { join } from "path"
 import { GardenModule } from "@garden-io/sdk/build/src/types"
 import { expect } from "chai"
 import { makeTestGarden, TestGarden } from "@garden-io/sdk/build/src/testing"
-import { gardenPlugin } from ".."
-import { JibBuildAction } from "../util"
+import { gardenPlugin } from "../src"
+import { JibBuildAction } from "../src/util"
 import { Resolved } from "@garden-io/core/build/src/actions/types"
 import { ResolvedConfigGraph } from "@garden-io/core/build/src/graph/config-graph"
 import { createActionLog } from "@garden-io/core/build/src/logger/log-entry"
