+++ plugins/jib/tsconfig.json
@@ -1,8 +1,8 @@
 {
   "extends": "../../tsconfig.base.json",
   "include": [
-    "*.ts",
-    "test/*.ts"
+    "src/*.ts",
+    "test/*.ts",
   ],
   "compilerOptions": {
     "composite": true,
