+++ plugins/pulumi/test/test-project-k8s/k8s-deployment/Pulumi.k8s-service.yaml
