+++ plugins/pulumi/test/test-project-k8s/k8s-deployment/package.json
