+++ plugins/pulumi/test/test-project-k8s/k8s-deployment/varfile1.yaml
