+++ plugins/pulumi/test/test-project-k8s/k8s-namespace/Pulumi.k8s-namespace.yaml
