+++ plugins/pulumi/test/test-project-k8s/k8s-namespace/index.ts
