+++ plugins/pulumi/test/test-project-k8s/tsconfig.json
