+++ plugins/pulumi/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.base.json",
   "include": [
-    "*.ts",
+    "src/*.ts",
     "test/*.ts"
   ],
   "compilerOptions": {
